research: notes — u64 note amounts (Note::WORDS 27->28)
SHRUGG units are 1e9 per coin; u32 cannot hold one coin
(shielded-pool spec §13). amount is now two words (lo, hi) in the
commitment preimage; input::* offsets shift accordingly (IN_AMOUNT ->
IN_AMOUNT_LO/HI, everything after it +1, input::COUNT 295 -> 296).
guests::transfer's NOTE_IN staging (both cm_in and cm_out blocks)
copies both amount words. viewing::Row.amount follows Note.amount to
u64 so scan()/verify_row() keep compiling and checking the real value
instead of a truncated one.

Adds a cheating test: a note whose real amount needs the high word
cannot be spent by a witness that zeroes it and only supplies the low
word — NOTE_COMMIT hashes both words, so the lie produces a cm_in
different from the tree's real leaf, and the guest's published output
digest (folded from the wrong root) can never match what `apply`
recomputes for the claimed anchor (BadDigest). At the old u32 width
there was no high word to lie about, so this amount could not even be
represented.

Re-measured the transfer guest (docs/06-viewing-keys.md): 1771->1783
instructions, 3896->3910 execution cycles, 190->192 execution
permutations (cm_in/cm_out's absorbed message crosses a rate-4 block
boundary, 28->29 words each), 443->446 digest rows, 4339->4356 total
cycles, 633->638 total permutations; gas tier unchanged at 14.

+++ research/src/viewing.rs
@@ -139,7 +139,7 @@ pub struct Row {
     pub role: Role,
     pub sender: Word8,
     pub receiver: Word8,
-    pub amount: u32,
+    pub amount: u64,
     pub asset: u32,
     pub time: u32,
     /// The created note's on-chain commitment.
